fix tests and add 1.5 fixtures
